separate out mysql storage types
